Revert "feat: add AddSliceDropdown component"
This reverts the previous commit 297912339e40e404b48c2a2f0d1b0464eac0518b.
