+++ packages/slice-machine/src/pages/slices.tsx
@@ -1,3 +1,11 @@
+import {
+  BackgroundIcon,
+  Button,
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@prismicio/editor-ui";
 import { SharedSlice as SharedSliceType } from "@prismicio/types-internal/lib/customtypes";
 import Head from "next/head";
 import { useRouter } from "next/router";
@@ -7,8 +15,8 @@ import { toast } from "react-toastify";
 import { BaseStyles, Flex, Link, Text } from "theme-ui";
 
 import { BreadcrumbItem } from "@/components/Breadcrumb";
+import { useAiSliceGenerationExperiment } from "@/features/builder/useAiSliceGenerationExperiment";
 import { CreateSliceFromImageModal } from "@/features/customTypes/customTypesBuilder/CreateSliceFromImageModal";
-import { AddSliceDropdown } from "@/features/slices/AddSliceDropdown";
 import { SharedSliceCard } from "@/features/slices/sliceCards/SharedSliceCard";
 import { SLICES_CONFIG } from "@/features/slices/slicesConfig";
 import { useScreenshotChangesModal } from "@/hooks/useScreenshotChangesModal";
@@ -37,6 +45,7 @@ import useSliceMachineActions from "@/modules/useSliceMachineActions";
 import { SliceMachineStoreType } from "@/redux/type";
 
 const SlicesIndex: React.FunctionComponent = () => {
+  const aiSliceGenerationExperiment = useAiSliceGenerationExperiment();
   const router = useRouter();
   const { modalPayload, onOpenModal } = useScreenshotChangesModal();
   const { openLoginModal } = useSliceMachineActions();
@@ -80,14 +89,6 @@ const SlicesIndex: React.FunctionComponent = () => {
     setIsRenameSliceModalOpen(true);
   };
 
-  const openCreateSliceModal = () => {
-    setIsCreateSliceModalOpen(true);
-  };
-
-  const closeCreateSliceModal = () => {
-    setIsCreateSliceModalOpen(false);
-  };
-
   const openCreateSliceFromImageModal = async () => {
     const isLoggedIn = await managerClient.user.checkIsLoggedIn();
 
@@ -113,10 +114,54 @@ const SlicesIndex: React.FunctionComponent = () => {
             <BreadcrumbItem>Slices</BreadcrumbItem>
           </AppLayoutBreadcrumb>
           {localLibraries?.length !== 0 && sliceCount !== 0 ? (
-            <AddSliceDropdown
-              onOpenCreateSliceFromImageModal={openCreateSliceFromImageModal}
-              onOpenCreateSliceModal={openCreateSliceModal}
-            />
+            <DropdownMenu>
+              <DropdownMenuTrigger>
+                <Button
+                  color="purple"
+                  startIcon="add"
+                  data-testid="add-new-slice-dropdown"
+                >
+                  Add
+                </Button>
+              </DropdownMenuTrigger>
+
+              <DropdownMenuContent align="end">
+                {aiSliceGenerationExperiment.eligible && (
+                  <DropdownMenuItem
+                    renderStartIcon={() => (
+                      <BackgroundIcon
+                        name="autoFixHigh"
+                        size="extraSmall"
+                        iconSize="small"
+                        radius={6}
+                        variant="solid"
+                        color="purple"
+                      />
+                    )}
+                    onSelect={() => void openCreateSliceFromImageModal()}
+                    description="Build a Slice based on your design image."
+                  >
+                    Generate from image
+                  </DropdownMenuItem>
+                )}
+                <DropdownMenuItem
+                  renderStartIcon={() => (
+                    <BackgroundIcon
+                      name="add"
+                      size="extraSmall"
+                      iconSize="small"
+                      radius={6}
+                      variant="solid"
+                      color="white"
+                    />
+                  )}
+                  onSelect={() => setIsCreateSliceModalOpen(true)}
+                  description="Build a custom Slice your way."
+                >
+                  Start from scratch
+                </DropdownMenuItem>
+              </DropdownMenuContent>
+            </DropdownMenu>
           ) : undefined}
         </AppLayoutHeader>
         <AppLayoutContent>
@@ -138,7 +183,9 @@ const SlicesIndex: React.FunctionComponent = () => {
                   >
                     <EmptyState
                       title={"What are Slices?"}
-                      onCreateNew={openCreateSliceModal}
+                      onCreateNew={() => {
+                        setIsCreateSliceModalOpen(true);
+                      }}
                       buttonText={"Create one"}
                       videoPublicIdUrl={VIDEO_WHAT_ARE_SLICES}
                       documentationComponent={
@@ -262,7 +309,9 @@ const SlicesIndex: React.FunctionComponent = () => {
                   }),
                 );
               }}
-              onClose={closeCreateSliceModal}
+              onClose={() => {
+                setIsCreateSliceModalOpen(false);
+              }}
             />
           )}
           <RenameSliceModal
